.backup/tool_service_backup_20251009_003423/requirements.txt
requests
beautifulsoup4
grpcio==1.62.1
grpcio-tools==1.62.1
grpcio-health-checking
protobuf==4.25.3
python-dotenv==1.0.1
grpcio-reflection
cachetools==5.3.3
selenium==4.18.1
selenium-wire==5.1.0
webdriver-manager==4.0.1
blinker
sympy